functions/stream_processor/requirements.txt
azure-functions
azure-servicebus
azure-storage-blob
python-dotenv
